# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ClanWolf DataGateway Backend API",
    "excerpt": "Documentation automatically generated by the <b>swagger-autogen</b> module.",
    "description": ""
  },
  "host": "dg.clanwolf.net",
  "basePath": "/",
  "schemes": "https",
  "paths": {
...
    "/c3attack/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/c3attack/{seasonId}": {
      "get": {
        "tags": [
          "C3 Attack"
        ],
        "summary": "Get all attacks of a given season",
        "description": "",
        "parameters": [
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the given season"
          },
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}